Add comments about shortname collisions
Using the shortname of the host has the downside of being unable to lock
two hosts with the same shortname and different domains.  However, since
the majority of users interact with shortname only, we're using
shortname to normalize and prevent the system from allowing a lock on
aixbuilder1 if aixbuilder1.delivery.puppetlabs.net is locked.