Fix race conidition for lock
In the previous implementation of lock_manager, there was a time of
check to time of use error in that a race could occur for a lock. This
commit fixes that by having the `lock` method use setnx (set if not
exists) from redis rather than a check for a key, then insert. This is
safer since it is an atomic operation.