(maint) Update rubocop configuration
Since rubocop renames, breaks, removes, and adds new cops every release,
it's kind of a pain. This commit updates to the lateset, 0.39.0, and
pins to that in the Gemfile. It also silences new issues the rubocop
found that were just fine a few weeks ago. *shakes fist*

I've pinned because otherwise a random version of rubocop can get pulled
in on TravisCI and cause tests to fail (as that happened this evening).