(MAINT) prune/fix outdated & unnecessary deps
We don't need to explicitly require 'json' -- it's a part of the Ruby
Stdlib. The only real exception is for Ruby 1.9.3, so we should make
its requirement a special exception instead of the rule.

And 'redis' is already provided by the Gemspec so we don't need that
for the development dependencies.